+++ src/EventStore.Core/Index/IMemTable.cs
@@ -5,7 +5,7 @@ namespace EventStore.Core.Index
     public interface IMemTable: ISearchTable
     {
         bool MarkForConversion();
-        void Add(ulong stream, int version, long position);
+        void Add(ulong stream, long version, long position);
         void AddEntries(IList<IndexEntry> entries);
     }
 }
