+++ src/EventStore.Core/Services/PersistentSubscription/IPersistentSubscriptionCheckpointReader.cs
@@ -4,6 +4,6 @@ namespace EventStore.Core.Services.PersistentSubscription
 {
     public interface IPersistentSubscriptionCheckpointReader
     {
-        void BeginLoadState(string subscriptionId, Action<int?> onStateLoaded);
+        void BeginLoadState(string subscriptionId, Action<long?> onStateLoaded);
     }
 }
