+++ src/EventStore.Projections.Core.Tests/Services/position_tagging/event_by_type_index_position_tagger/when_updating_event_by_type_index_position_tracker_from_a_tag.cs
@@ -19,7 +19,7 @@ namespace EventStore.Projections.Core.Tests.Services.position_tagging.event_by_t
             var tagger = new EventByTypeIndexPositionTagger(0, new[] {"type1", "type2"});
             var tracker = new PositionTracker(tagger);
 
-            var newTag = CheckpointTag.FromEventTypeIndexPositions(0, new TFPos(10, 5), new Dictionary<string, int> {{"type1", 1}, {"type2", 2}});
+            var newTag = CheckpointTag.FromEventTypeIndexPositions(0, new TFPos(10, 5), new Dictionary<string, long> {{"type1", 1}, {"type2", 2}});
 
             tracker.UpdateByCheckpointTagInitial(newTag);
             _tag = tracker.LastTag;
