+++ src/EventStore.Projections.Core/Services/Management/ManagedProjection.cs
@@ -49,8 +49,8 @@ namespace EventStore.Projections.Core.Services.Management
             public bool? CreateTempStreams { get; set; }
             public bool? CheckpointsDisabled { get; set; }
             public bool? TrackEmittedStreams { get; set; }
-            public int? Epoch { get; set; }
-            public int? Version { get; set; }
+            public long? Epoch { get; set; }
+            public long? Version { get; set; }
             public SerializedRunAs RunAs { get; set; }
         }
 
@@ -82,7 +82,7 @@ namespace EventStore.Projections.Core.Services.Management
         private readonly ITimeProvider _timeProvider;
         private readonly Guid _workerId;
         private readonly Guid _id;
-        private readonly int _projectionId;
+        private readonly long _projectionId;
         private readonly string _name;
         private readonly bool _enabledToRun;
         private ManagedProjectionState _state;
@@ -93,7 +93,7 @@ namespace EventStore.Projections.Core.Services.Management
         //private List<IEnvelope> _debugStateRequests;
         private ProjectionStatistics _lastReceivedStatistics;
         private DateTime _lastAccessed;
-        private int _lastWrittenVersion = -1;
+        private long _lastWrittenVersion = -1;
         private IPrincipal _runAs;
         //TODO: slave (extract into derived class)
 
@@ -113,7 +113,7 @@ namespace EventStore.Projections.Core.Services.Management
         public ManagedProjection(
             Guid workerId,
             Guid id,
-            int projectionId,
+            long projectionId,
             string name,
             bool enabledToRun,
             ILogger logger,
